Show NVIDIA Smooth Motion overlay status

+++ hook/common/fg_detection.h
@@ -78,7 +78,7 @@ public:
         if (mult >= 2) {
             float output = cachedOutputFPS.load();
             if (output > 1.0f)
-                return output / mult;
+                return output / static_cast<float>(mult);
         }
         return cachedBaseFPS.load();
     }
